refactor: sortableCards into its own component
